chore_: `SendType` moved to another file
